Rename RepoToken -> AppToken
